chore: move ethfiltererror
