+++ crates/rpc/rpc-eth-types/src/lib.rs
@@ -36,7 +36,6 @@ pub use gas_oracle::{
     GasCap, GasPriceOracle, GasPriceOracleConfig, GasPriceOracleResult, RPC_DEFAULT_GAS_CAP,
 };
 pub use id_provider::EthSubscriptionIdProvider;
-pub use logs_utils::EthFilterError;
 pub use pending_block::{PendingBlock, PendingBlockEnv, PendingBlockEnvOrigin};
 pub use receipt::ReceiptBuilder;
 pub use transaction::TransactionSource;
